user-service/requirements.txt
#uvicorn app.main:app --reload
fastapi==0.116.1
pydantic==2.11.7
passlib==1.7.4
pybloom_live==4.0.0
bitarray==3.5.2
validate_email==1.3
snowflake-id==1.0.2
cryptography==3.4.8
bcrypt==4.3.0
uvicorn[standard]
email-validator
aiomysql==0.2.0
python-dotenv